+++ r4/src/test/java/io/github/jy95/fds/r4/translators/RateRatioTest.java
@@ -1,47 +1,31 @@
 package io.github.jy95.fds.r4.translators;
 
-import io.github.jy95.fds.r4.DosageAPIR4;
-import io.github.jy95.fds.r4.AbstractFhirTest;
 import io.github.jy95.fds.common.types.DisplayOrder;
+import io.github.jy95.fds.common.types.DosageAPI;
+import io.github.jy95.fds.r4.DosageAPIR4;
+import io.github.jy95.fds.r4.config.FDSConfigR4;
+import io.github.jy95.fds.translators.AbstractRateRatioTest;
 import org.hl7.fhir.r4.model.Dosage;
 import org.hl7.fhir.r4.model.Quantity;
 import org.hl7.fhir.r4.model.Ratio;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.MethodSource;
 
 import java.util.List;
 import java.util.Locale;
-import java.util.concurrent.ExecutionException;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-public class RateRatioTest extends AbstractFhirTest {
-
-    @ParameterizedTest
-    @MethodSource("localeProvider")
-    void testNoRateRatio(Locale locale) throws ExecutionException, InterruptedException {
-        Dosage dosage = new Dosage();
-        DosageAPIR4 dosageUtils = DayOfWeekTest.getDosageAPI(locale, DisplayOrder.RATE_RATIO);
-        String result = dosageUtils.asHumanReadableText(dosage).get();
-        assertEquals("", result);
-    }
-
-    @ParameterizedTest
-    @MethodSource("localeProvider")
-    void testEmptyRateRatio(Locale locale) throws ExecutionException, InterruptedException {
+public class RateRatioTest extends AbstractRateRatioTest<FDSConfigR4, Dosage> {
+    
+    @Override
+    protected Dosage generateEmptyRateRatio() {
         Dosage dosage = new Dosage();
         Ratio ratio = new Ratio();
         Dosage.DosageDoseAndRateComponent doseAndRateComponent1 = new Dosage.DosageDoseAndRateComponent();
         doseAndRateComponent1.setRate(ratio);
         dosage.setDoseAndRate(List.of(doseAndRateComponent1));
-        DosageAPIR4 dosageUtils = DayOfWeekTest.getDosageAPI(locale, DisplayOrder.RATE_RATIO);
-        String result = dosageUtils.asHumanReadableText(dosage).get();
-        assertEquals("", result);
+        return dosage;
     }
 
-    @ParameterizedTest
-    @MethodSource("localeProvider")
-    void testOnlyNumerator(Locale locale) throws ExecutionException, InterruptedException {
+    @Override
+    protected Dosage generateOnlyNumerator() {
         Dosage dosage = new Dosage();
         Ratio ratio = new Ratio();
         Quantity numerator = new Quantity(10);
@@ -50,15 +34,11 @@ public class RateRatioTest extends AbstractFhirTest {
         Dosage.DosageDoseAndRateComponent doseAndRateComponent1 = new Dosage.DosageDoseAndRateComponent();
         doseAndRateComponent1.setRate(ratio);
         dosage.setDoseAndRate(List.of(doseAndRateComponent1));
-        DosageAPIR4 dosageUtils = DayOfWeekTest.getDosageAPI(locale, DisplayOrder.RATE_RATIO);
-        String result = dosageUtils.asHumanReadableText(dosage).get();
-        String expected = getExpectedText1(locale);
-        assertEquals(expected, result);
+        return dosage;
     }
 
-    @ParameterizedTest
-    @MethodSource("localeProvider")
-    void testOnlyDenominator(Locale locale) throws ExecutionException, InterruptedException {
+    @Override
+    protected Dosage generateOnlyDenominator() {
         Dosage dosage = new Dosage();
         Ratio ratio = new Ratio();
         Quantity denominator = new Quantity(10);
@@ -67,15 +47,11 @@ public class RateRatioTest extends AbstractFhirTest {
         Dosage.DosageDoseAndRateComponent doseAndRateComponent1 = new Dosage.DosageDoseAndRateComponent();
         doseAndRateComponent1.setRate(ratio);
         dosage.setDoseAndRate(List.of(doseAndRateComponent1));
-        DosageAPIR4 dosageUtils = DayOfWeekTest.getDosageAPI(locale, DisplayOrder.RATE_RATIO);
-        String result = dosageUtils.asHumanReadableText(dosage).get();
-        String expected = getExpectedText2(locale);
-        assertEquals(expected, result);
+        return dosage;
     }
 
-    @ParameterizedTest
-    @MethodSource("localeProvider")
-    void testNumeratorAndDenominator(Locale locale) throws ExecutionException, InterruptedException {
+    @Override
+    protected Dosage generateNumeratorAndDenominator() {
         Dosage dosage = new Dosage();
         Ratio ratio = new Ratio();
         Quantity numerator = new Quantity(10);
@@ -87,15 +63,11 @@ public class RateRatioTest extends AbstractFhirTest {
         Dosage.DosageDoseAndRateComponent doseAndRateComponent1 = new Dosage.DosageDoseAndRateComponent();
         doseAndRateComponent1.setRate(ratio);
         dosage.setDoseAndRate(List.of(doseAndRateComponent1));
-        DosageAPIR4 dosageUtils = DayOfWeekTest.getDosageAPI(locale, DisplayOrder.RATE_RATIO);
-        String result = dosageUtils.asHumanReadableText(dosage).get();
-        String expected = getExpectedText3(locale);
-        assertEquals(expected, result);
+        return dosage;
     }
 
-    @ParameterizedTest
-    @MethodSource("localeProvider")
-    void testTiterCase(Locale locale) throws ExecutionException, InterruptedException {
+    @Override
+    protected Dosage generateTiterCase() {
         Dosage dosage = new Dosage();
         Ratio ratio = new Ratio();
         Quantity numerator = new Quantity(1);
@@ -105,15 +77,11 @@ public class RateRatioTest extends AbstractFhirTest {
         Dosage.DosageDoseAndRateComponent doseAndRateComponent1 = new Dosage.DosageDoseAndRateComponent();
         doseAndRateComponent1.setRate(ratio);
         dosage.setDoseAndRate(List.of(doseAndRateComponent1));
-        DosageAPIR4 dosageUtils = DayOfWeekTest.getDosageAPI(locale, DisplayOrder.RATE_RATIO);
-        String result = dosageUtils.asHumanReadableText(dosage).get();
-        String expected = getExpectedText4(locale);
-        assertEquals(expected, result);
+        return dosage;
     }
 
-    @ParameterizedTest
-    @MethodSource("localeProvider")
-    void testCommon(Locale locale) throws ExecutionException, InterruptedException {
+    @Override
+    protected Dosage generateCommonCase() {
         Dosage dosage = new Dosage();
         Ratio ratio = new Ratio();
         Quantity numerator = new Quantity(10);
@@ -125,70 +93,24 @@ public class RateRatioTest extends AbstractFhirTest {
         Dosage.DosageDoseAndRateComponent doseAndRateComponent1 = new Dosage.DosageDoseAndRateComponent();
         doseAndRateComponent1.setRate(ratio);
         dosage.setDoseAndRate(List.of(doseAndRateComponent1));
-        DosageAPIR4 dosageUtils = DayOfWeekTest.getDosageAPI(locale, DisplayOrder.RATE_RATIO);
-        String result = dosageUtils.asHumanReadableText(dosage).get();
-        String expected = getExpectedText5(locale);
-        assertEquals(expected, result);
+        return dosage;
     }
 
-    private String getExpectedText1(Locale locale) {
-        if (locale.equals(Locale.ENGLISH)) {
-            return "at a rate of 10 ml";
-        } else if (locale.equals(Locale.FRENCH)) {
-            return "au taux de 10 ml";
-        } else if (locale.equals(Locale.GERMAN)) {
-            return "mit einem Verhältnis von 10 ml";
-        } else {
-            return "met een verhouding van 10 ml";
-        }
+    @Override
+    public DosageAPI<FDSConfigR4, Dosage> getDosageAPI(Locale locale, DisplayOrder displayOrder) {
+        return new DosageAPIR4(FDSConfigR4.builder()
+                .displayOrder(List.of(displayOrder))
+                .locale(locale)
+                .build());
     }
 
-    private String getExpectedText2(Locale locale) {
-        if (locale.equals(Locale.ENGLISH)) {
-            return "at a rate of 10 ml";
-        } else if (locale.equals(Locale.FRENCH)) {
-            return "au taux de 10 ml";
-        } else if (locale.equals(Locale.GERMAN)) {
-            return "mit einem Verhältnis von 10 ml";
-        } else {
-            return "met een verhouding van 10 ml";
-        }
+    @Override
+    public DosageAPI<FDSConfigR4, Dosage> getDosageAPI(FDSConfigR4 config) {
+        return new DosageAPIR4(config);
     }
 
-    private String getExpectedText3(Locale locale) {
-        if (locale.equals(Locale.ENGLISH)) {
-            return "at a rate of 10 mg per ml";
-        } else if (locale.equals(Locale.FRENCH)) {
-            return "au taux de 10 mg par ml";
-        } else if (locale.equals(Locale.GERMAN)) {
-            return "mit einem Verhältnis von 10 mg pro ml";
-        } else {
-            return "met een verhouding van 10 mg per ml";
-        }
+    @Override
+    public Dosage generateEmptyDosage() {
+        return new Dosage();
     }
-
-    private String getExpectedText4(Locale locale) {
-        if (locale.equals(Locale.ENGLISH)) {
-            return "at a rate of 1 : 128";
-        } else if (locale.equals(Locale.FRENCH)) {
-            return "au taux de 1 : 128";
-        } else if (locale.equals(Locale.GERMAN)) {
-            return "mit einem Verhältnis von 1 : 128";
-        } else {
-            return "met een verhouding van 1 : 128";
-        }
-    }
-
-    private String getExpectedText5(Locale locale) {
-        if (locale.equals(Locale.ENGLISH)) {
-            return "at a rate of 10 mg every 2 ml";
-        } else if (locale.equals(Locale.FRENCH)) {
-            return "au taux de 10 mg chaque 2 ml";
-        } else if (locale.equals(Locale.GERMAN)) {
-            return "mit einem Verhältnis von 10 mg jeder 2 ml";
-        } else {
-            return "met een verhouding van 10 mg elke 2 ml";
-        }
-    }
-
 }
